bot/requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
Pillow==10.2.0
aiofiles==23.2.1python.